Fix token auth. Fix games sync. Fix game launch. Add live replays. Add custom UI for game cards

+++ FAF.Domain.LobbyServer/Team.cs
@@ -5,7 +5,7 @@ namespace FAF.Domain.LobbyServer
     public class TeamIds
     {
         [JsonPropertyName("team_id")]
-        public long TeamId { get; set; }
+        public int TeamId { get; set; }
 
         [JsonPropertyName("player_ids")]
         public long[] PlayerIds { get; set; }
